Added Twitter Bootstrap Local Backup
Added local version of twitter bootstrap css and js along with logic to 
load them if the CDN fails for any layout.

Needed to remove a code block from _includes/jssor-slideshow-js.html 
that reloads the page if the window width changes since that results in 
a refresh loop from the local CSS backup being loaded.